Modified Teller, Port, and Managers
The teller and port classes were changed to be threads and that share a common pair of queues. They access the queues synchronously and if queue is empty call wait on the queue. Additionally implement proper usage of locks when accessing the data repository and hash entries. Created a rollback system to undo changes and reattempt locking entire data repository instead. After completion, changes are committed to persistent storage and the log file finalized. Still need to add replicate system.